requirements.txt
asgiref==3.3.1
Django==3.1.5
djangorestframework==3.12.2
pytz==2020.5
redis==3.5.3
sqlparse==0.4.1
